Put effects and configuration to separate modules.
Also reenable support for Plasma 2040W.
